feat(citrea_client): Implement get_citrea_l2_height_range for the Citrea client and make it's return value inclusive.
